refactor: move gh api interactions to gh provider
