+++ lua/gh-actions/health.lua
@@ -1,46 +1,33 @@
 local M = {}
 
 function M.check()
-  local health = vim.health or require('health')
+  local health = vim.health
 
-  local start = health.start or health.report_start
-  local ok = health.ok or health.report_ok
-  local warn = health.warn or health.report_warn
-  local error = health.error or health.report_error
-
-  start('Checking ability to parse yaml files')
+  health.start('Checking ability to parse yaml files')
 
   local has_rust_module = pcall(require, 'gh-actions.rust')
 
   if has_rust_module then
-    ok('Found rust module')
+    health.ok('Found rust module')
   else
-    warn('No rust module found')
+    health.warn('No rust module found')
   end
 
   local has_yq_installed = vim.fn.executable('yq') == 1
 
   if has_yq_installed then
-    ok('Found yq executable')
+    health.ok('Found yq executable')
   else
-    warn('No yq executable found')
+    health.warn('No yq executable found')
   end
 
   if has_rust_module or has_yq_installed then
-    ok('Found yaml parser')
+    health.ok('Found yaml parser')
   else
-    error('No yaml parser found')
+    health.error('No yaml parser found')
   end
 
-  start('Checking for GitHub token')
-
-  local k, token = pcall(require('gh-actions.github').get_github_token)
-
-  if k and token then
-    ok('Found GitHub token')
-  else
-    error('No GitHub token found')
-  end
+  require('gh-actions.providers.github.rest.health').check()
 end
 
 return M
