+++ lua/lualine/components/gh-actions.lua
@@ -1,5 +1,5 @@
 local function gh()
-  return require('gh-actions.github')
+  return require('gh-actions.providers.github.rest._api')
 end
 
 local component = require('lualine.component'):extend()
@@ -36,15 +36,15 @@ function component:update_status()
   local state = self.store.get_state()
 
   local latest_workflow_run = state.workflow_runs and state.workflow_runs[1]
-    or {}
+      or {}
 
   if not latest_workflow_run.status then
     return ''
   end
 
   return self.icons.get_workflow_run_icon(latest_workflow_run)
-    .. ' '
-    .. latest_workflow_run.name
+      .. ' '
+      .. latest_workflow_run.name
 end
 
 return component
